move recipes and complexity into callbacks and components
